% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertAndRetrieveTidyValue.R
\name{assertAndRetrieveTidyValue}
\alias{assertAndRetrieveTidyValue}
\title{Retriever of `tidyselect`}
\usage{
assertAndRetrieveTidyValue(
  x,
  value,
  name = deparse(substitute(value)),
  optional = FALSE
)
}
\arguments{
\item{x}{The data with the potential value}

\item{value}{The value}

\item{name}{The name of the value}

\item{optional}{Is the value optional}
}
\value{
value with attribute
}
\description{
As forestpot has evolved we now primarily use `tidyverse` select style. This
function helps with backward compatibility
}
