% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_beta}
\alias{demo_beta}
\title{Linear Associations of NMR-quantified Biomarkers to BMI}
\format{A data frame (tibble) with 228 rows and 2 columns:
\describe{
  \item{abbrev}{Biomarker abbreviation}
  \item{cohort1}{Association values for simulated cohort 1}
  \item{cohort2}{Association values for simulated cohort 2}
}}
\usage{
demo_beta
}
\description{
A dataframe containing linear associations of NMR-quantified biomarkers to BMI as estimated using simulated data.
}
\keyword{datasets}
