% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_pval}
\alias{demo_pval}
\title{P-values of Linear Associations of NMR-quantified Biomarkers to BMI}
\format{A data frame (tibble) with 228 rows and 2 columns:
\describe{
  \item{abbrev}{Biomarker abbreviation}
  \item{cohort1}{p-values for simulated cohort 1}
  \item{cohort2}{p-values for simulated cohort 2}
}}
\usage{
demo_pval
}
\description{
A dataframe containing p-values for linear associations of NMR-quantified biomarkers to BMI as estimated using simulated data. p-values correspond to the demo_beta and demo_se dataframe values
}
\keyword{datasets}
