% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_se}
\alias{demo_se}
\title{Standard Error Values of Linear Associations of NMR-quantified Biomarkers to BMI}
\format{A data frame (tibble) with 228 rows and 2 columns:
\describe{
  \item{abbrev}{Biomarker abbreviation}
  \item{cohort1}{Std. error values for simulated cohort 1}
  \item{cohort2}{Std. error values for simulated cohort 2}
}}
\usage{
demo_se
}
\description{
A dataframe containing standard error values for linear associations of NMR-quantified biomarkers to BMI as estimated using simulated data. Std. error values correspond to the demo_beta dataframe values
}
\keyword{datasets}
