% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_drawLines.R
\name{drawHorizontalLines}
\alias{drawHorizontalLines}
\title{Draws the horizontal lines}
\usage{
drawHorizontalLines(lines, number_of_rows, colwidths, graph.pos)
}
\arguments{
\item{number_of_rows}{Number of rows}

\item{colwidths}{Vector with column widths}

\item{graph.pos}{The position of the graph element within the table of text. The
position can be \code{1-(ncol(labeltext) + 1)}. You can also choose set the position
to \code{"left"} or \code{"right"}.}
}
\description{
Draws the horizontal lines
}
\keyword{internal}
