% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot-package.R
\docType{package}
\name{forestplot-package}
\alias{forestplot-package}
\title{The forest plot function, \code{\link{forestplot2}}, is a more general
version of the original \pkg{rmeta}-packages \code{\link[rmeta]{forestplot}}
implementation. The aim is at using forest plots for more than
just meta-analyses.}
\description{
The forest plot function, \code{\link{forestplot2}}, is a
more general version of the original \pkg{rmeta}-packages
\code{\link[rmeta]{forestplot}} implementation. The aim is
at using forest plots for more than just meta-analyses.
}

