% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestplot_helpers.R
\name{fpTxtGp}
\alias{fpTxtGp}
\title{Get font settings for forestplot}
\usage{
fpTxtGp(label, summary, xlab, title, ticks, legend, legend.title, cex = 1)
}
\arguments{
  \item{label}{The text labels}

  \item{summary}{The summary labels}

  \item{xlab}{The xlab text}

  \item{title}{The plot title}

  \item{ticks}{The ticks associated with the xlab}

  \item{legend}{The legend text}

  \item{legend.title}{The legend title}

  \item{cex}{The font size}
}
\value{
A list of the \code{fpTxtGp} class
}
\description{
This function autogenerates all the
\code{\link[grid]{gpar}()} elements for the different text
elements within the graph.
}
\examples{
fpTxtGp(label=gpar(fontfamily="HersheySerif"))
}

