% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_helpers.R
\name{prDefaultGp}
\alias{prDefaultGp}
\title{Construct default parameters from arguments that may include missing arguments}
\usage{
prDefaultGp(col, lwd, lty)
}
\arguments{
\item{col}{Line color (or missing)}

\item{lwd}{Line width (or missing)}

\item{lty}{Line type (or missing)}
}
\value{
a \code{\link[grid]{gpar}} object
containing these three attributes
}
\description{
Construct default parameters from arguments that may include missing arguments
}
