% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_prepAlign.R
\name{prepAlign}
\alias{prepAlign}
\title{Prepares graph position}
\usage{
prepAlign(align, graph.pos, nc)
}
\arguments{
\item{align}{Vector giving alignment (l,r,c) for the table columns}

\item{graph.pos}{An integer indicating the position of the graph}

\item{nc}{The number of columns}
}
\value{
Returns vector of `"l", "c", "r"` values
}
\description{
Prepares the graph position so that it matches the label size
}
