% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_prepGraphPositions.R
\name{prepGraphPositions}
\alias{prepGraphPositions}
\title{Prepares graph position}
\usage{
prepGraphPositions(graph.pos, nc)
}
\arguments{
\item{graph.pos}{The position of the graph element within the table of text. The
position can be \code{1-(ncol(labeltext) + 1)}. You can also choose set the position
to \code{"left"} or \code{"right"}.}

\item{nc}{The number of columns}
}
\value{
Returns number indicating the graph position
}
\description{
Prepares the graph position so that it matches the label size
}
