% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGridMargins.R
\name{prepGridMargins}
\alias{prepGridMargins}
\title{Convert margins to viewport npc margins}
\usage{
prepGridMargins(mar)
}
\arguments{
\item{mar}{A vector of margins, at positions:
- 1 = bottom
- 2 = left
- 3 = top
- 4 = right}
}
\value{
Returns a list with `bottom`, `left`, `top`, and `right` as `unit("npc")`
}
\description{
Convert margins to viewport npc margins
}
