% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_prepLines.R
\name{prepLines}
\alias{prepLines}
\title{Prepares the lines for the plot}
\usage{
prepLines(lines, is.summary, number_of_rows, number_of_columns, col, shapes_gp)
}
\arguments{
\item{is.summary}{A vector indicating by \code{TRUE}/\code{FALSE} if
the value is a summary value which means that it will have a different
font-style}

\item{number_of_rows}{Total number of rows}

\item{number_of_columns}{Total number of columns}

\item{col}{Set the colors for all the elements. See \code{\link{fpColors}} for
details}

\item{shapes_gp}{Sets graphical parameters (squares and lines widths, styles, etc.)
of all shapes drawn (squares, lines, diamonds, etc.). This overrides \code{col},
\code{lwd.xaxis}, \code{lwd.zero}, \code{lwd.ci} and \code{lty.ci}.}
}
\description{
Prepares the lines for the plot
}
\keyword{internal}
