% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/private.R
\name{prFpDrawLines}
\alias{prFpDrawLines}
\title{Draws the horizontal lines}
\usage{
prFpDrawLines(hrzl_lines, nr, colwidths)
}
\arguments{
  \item{hrzl_lines}{Add horizontal lines to graph. Can
  either be \code{TRUE} or a \code{list} of
  \code{\link[grid]{gpar}}. See line section below for
  details.}

  \item{nr}{Number of rows}

  \item{colwidths}{Vector with column widths}
}
\description{
Draws the horizontal lines
}
\keyword{internal}

