% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{prFpGetConfintFnList}
\alias{prFpGetConfintFnList}
\title{Get a function list}
\usage{
prFpGetConfintFnList(fn, no_rows, no_depth, missing_rows, is.summary, summary)
}
\arguments{
\item{fn}{The function list/matrix. If a list it
should be in the format [[row]][[col]], the function
tries to handle this but in cases where the columns
and rows are the same it will not know what is a column
and what is a row.}

\item{no_rows}{Number of rows}

\item{no_depth}{Number of columns}

\item{missing_rows}{The rows that don't have a CI}

\item{is.summary}{A vector indicating by \code{TRUE}/\code{FALSE} if
the value is a summary value which means that it will have a different
font-style}
}
\value{
\code{list} The function returns a list that has
the format [[row]][[col]] where each element contains the
function that you need to call using the \code{\link[base]{as.call}}
and \code{\link[base]{eval}} functions: \code{eval(as.call(list(fn[[row]][[col]], arg_1 = 1, arg_2 = 2)))}
}
\description{
This function helps the \code{\link{forestplot}}
to deal with multiple drawing functions for the
confidence intervals.
}
\keyword{internal}
