% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{prFpGetLines}
\alias{prFpGetLines}
\title{Prepares the hrzl_lines for the plot}
\usage{
prFpGetLines(
  hrzl_lines = NULL,
  is.summary,
  total_columns,
  col,
  shapes_gp = fpShapesGp()
)
}
\arguments{
\item{hrzl_lines}{Add horizontal lines to graph. Can either be \code{TRUE} or a \code{list}
of \code{\link[grid]{gpar}}. See line section below for details.}

\item{is.summary}{A vector indicating by \code{TRUE}/\code{FALSE} if
the value is a summary value which means that it will have a different
font-style}

\item{total_columns}{Total number of columns}

\item{col}{Set the colors for all the elements. See \code{\link{fpColors}} for
details}

\item{shapes_gp}{Sets graphical parameters (squares and lines widths, styles, etc.)
of all shapes drawn (squares, lines, diamonds, etc.). This overrides \code{col},
\code{lwd.xaxis}, \code{lwd.zero}, \code{lwd.ci} and \code{lty.ci}.}
}
\description{
Prepares the hrzl_lines for the plot
}
\keyword{internal}
