% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{prFpPrepareLegendMarker}
\alias{prFpPrepareLegendMarker}
\title{Prepares the legend marker function}
\usage{
prFpPrepareLegendMarker(fn.legend, col_no, fn.ci_norm)
}
\arguments{
\item{fn.legend}{The unknown parameter}

\item{col_no}{The number of columns}

\item{fn.ci_norm}{The original fn.ci_norm input}
}
\value{
\code{list}
}
\description{
Prepares the legend marker function
}
\keyword{internal}

