% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{prFpPrintLabels}
\alias{prFpPrintLabels}
\title{Plots the labels}
\usage{
prFpPrintLabels(labels, nc, nr, graph.pos)
}
\arguments{
\item{labels}{A list to the labels}

\item{nc}{Number of columns}

\item{nr}{Number of rows}

\item{graph.pos}{The position of the graph element within the table of text. The
position can be \code{1-(ncol(labeltext) + 1)}. You can also choose set the positin
to \code{"left"} or \code{"right"}.}
}
\value{
\code{void}
}
\description{
This is a helper function to the \code{\link{forestplot}}
function.
}
\keyword{internal}
