% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_prGetLabelsList.R
\name{prGetLabelsList}
\alias{prGetLabelsList}
\title{Gets the forestplot labels}
\usage{
prGetLabelsList(labels, align, is.summary, txt_gp, col)
}
\arguments{
\item{labels}{A `forestplot_labeltext` object}

\item{align}{Alignment, should be equal to \code{attr(labels, "no_cols")}}

\item{is.summary}{A vector indicating by \code{TRUE}/\code{FALSE} if
the value is a summary value which means that it will have a different
font-style}

\item{txt_gp}{Set the fonts etc for all text elements. See \code{\link{fpTxtGp}}
for details}

\item{col}{Set the colors for all the elements. See \code{\link{fpColors}} for
details}
}
\value{
\code{list} A list with \code{attr(labels, "no_cols")} where each element contains
 a list of \code{attr(labels, "no_rows")} elements with attributes width/height for each
 element and max_width/max_height for the total
}
\description{
A function that gets all the labels
}
\keyword{internal}
