% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/private.R
\name{prGparMerge}
\alias{prGparMerge}
\title{Merges two \code{\link[grid]{gpar}} elements}
\usage{
prGparMerge(l1, l2)
}
\arguments{
  \item{l1}{A \code{\link[grid]{gpar}} element}

  \item{l2}{A \code{\link[grid]{gpar}} element}
}
\value{
Returns a \code{\link[grid]{gpar}} element
}
\description{
The second elements overrides any conflicting elements
within the first
}
\keyword{internal}

