% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private.R
\name{prGridPlotTitle}
\alias{prGridPlotTitle}
\title{Adds a title to the plot}
\usage{
prGridPlotTitle(title, gp, space_below)
}
\arguments{
  \item{title}{The title as accepted by \code{\link[grid]
  {textGrob}}} \item{space_below}{The space below, defaults
  to 1/5 of the title height}
}
\value{
\code{NULL} The function does not return a value
}
\description{
Adds the title and generates a new main viewport below the
title
}
\keyword{internal}

