% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_helpers.R
\name{prMergeGp}
\alias{prMergeGp}
\title{A function to merge two sets of graphical parameters}
\usage{
prMergeGp(weak = gpar(), strong = gpar())
}
\arguments{
\item{weak}{A \code{\link[grid]{gpar}}}

\item{strong}{Another \code{\link[grid]{gpar}}, with parameters taking precedence
over \code{weak}}
}
\value{
A \code{\link[grid]{gpar}} merging attributes of both \code{weak} and \code{strong}
}
\description{
A function to merge two sets of graphical parameters
}
