% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeLoadPackage.R
\name{safeLoadPackage}
\alias{safeLoadPackage}
\title{Safely loads package}
\usage{
safeLoadPackage(package)
}
\arguments{
\item{package}{string naming the package/name space to load.}
}
\description{
Stops if the package doesn't exist
}
