% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomarkers}
\alias{biomarkers}
\title{NMR-quantified Biomarker Names and Abbreviations}
\format{A data frame (tibble) with 228 rows and 4 columns:
\describe{
  \item{abbrev}{Biomarker abbreviation}
  \item{fullname}{Biomarker full name}
  \item{forest_plot_disp_name}{Biomarker display name (for forest plot purposes)}
  \item{forest_plot_categories}{Suggestive biomarker category (for forestplot purposes)}
}}
\usage{
biomarkers
}
\description{
A dataframe containing the abbreviations and spelled-out names of 228 NMR-quantified biomarkers.
}
\keyword{datasets}
