% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-HRQoL.R
\docType{data}
\name{dfHRQoL}
\alias{dfHRQoL}
\title{Regression coefficients and confidence intervals from HRQoL study}
\description{
The data is a dataframe with the Swedish and the Danish coefficients
for health related quality of life (HRQoL) 1 year after total hip arthroplasty surgery.
The age is modeled as a spline and is therefore presented as a contrast.
}
\author{
Max Gordon \email{max@gforge.se}
}
\keyword{data}
