% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fpcolors}
\alias{fpcolors}
\title{A Palette of 5 Colors}
\format{An object of class \code{character} of length 5.}
\usage{
fpcolors
}
\description{
A named vector containing the hex numbers of 5 colors.
}
\keyword{datasets}
