% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data-HRQoL.R
\docType{data}
\name{HRQoL}
\alias{HRQoL}
\title{Coefficients and confidence interval for HRQoL in Sweden}
\description{
The data is a list containing the Swedish and the Danish
coefficients for health related quality of life 1 year
after surgery. The age is modelled as a spline and is
therefore presented as a contrast.
}
\author{
Max Gordon \email{max@gforge.se}
}
\keyword{data}

