% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-HRQoL.R
\docType{data}
\name{HRQoL}
\alias{HRQoL}
\title{Regression coefficients and confidence intervals from HRQoL study}
\description{
The data is a list containing the Swedish and the Danish
coefficients for health related quality of life (HRQoL)
1 year after total hip arthroplasty surgery. The age
is modelled as a spline and is therefore presented as a
contrast.
}
\author{
Max Gordon \email{max@gforge.se}
}
\keyword{data}

