% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGridMargins.R
\name{prepGridMargins}
\alias{prepGridMargins}
\title{Convert margins to viewport npc margins}
\usage{
prepGridMargins(mar)
}
\arguments{
\item{mar}{A vector of margins, at positions:
\itemize{
\item 1 = bottom
\item 2 = left
\item 3 = top
\item 4 = right
}}
}
\value{
Returns a list with \code{bottom}, \code{left}, \code{top}, and \code{right} as \code{unit("npc")}
}
\description{
Convert margins to viewport npc margins
}
