% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private.R
\name{prFpPrintXaxis}
\alias{prFpPrintXaxis}
\title{Plots the x-axis for forestplot}
\usage{
prFpPrintXaxis(axisList, col, lwd.zero)
}
\arguments{
\item{axisList}{The list from \code{\link{prFpGetGraphTicksAndClips}}}

\item{col}{Set the colors for all the elements. See \code{\link{fpColors}} for
details}

\item{lwd.zero}{lwd for the vertical line that gives the no-effect line, see \code{\link[grid]{gpar}}}
}
\value{
void
}
\description{
A helper function to the \code{\link{forestplot}}
function.
}
\keyword{internal}

