% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{prListRep}
\alias{prListRep}
\title{An alternative to rep()}
\usage{
prListRep(x, length.out)
}
\arguments{
\item{x}{The list to be repeated}

\item{length.out}{The length of the resulting list}
}
\value{
\code{list}
}
\description{
The rep() doesn't work with length.out
when lists are supposed to be their own
elements
}
\keyword{internal}
