% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{prPopulateList}
\alias{prPopulateList}
\title{Populate a list corresponding to matrix specs}
\usage{
prPopulateList(elmnt, no_rows, no_depth, missing_rows, is.summary, summary)
}
\arguments{
\item{elmnt}{The element item/list/matrix. If a list it
should be in the format [\link{row}][\link{col}], the function
tries to handle this but in cases where the columns
and rows are the same it will not know what is a column
and what is a row.}

\item{no_rows}{Number of rows}

\item{no_depth}{Number of outcomes per row, i.e. depth}

\item{missing_rows}{The rows that don't have data}

\item{is.summary}{A vector indicating by \code{TRUE}/\code{FALSE} if
the value is a summary value which means that it will have a different
font-style}
}
\value{
\code{list} The function returns a list that has
the format [\link{row}][\link{col}] where each element contains the
corresponding element
}
\description{
This function helps the \code{\link{forestplot}}
to deal with different arguments for the
confidence intervals.
}
\keyword{internal}
