% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_styling.R
\name{fp_txt_italic}
\alias{fp_txt_italic}
\alias{fp_txt_bold}
\alias{fp_txt_plain}
\alias{fp_txt_gp}
\alias{fp_align_left}
\alias{fp_align_center}
\alias{fp_align_right}
\title{Text styling}
\usage{
fp_txt_italic(txt)

fp_txt_bold(txt)

fp_txt_plain(txt)

fp_txt_gp(txt, gp)

fp_align_left(txt)

fp_align_center(txt)

fp_align_right(txt)
}
\arguments{
\item{txt}{The text to styl}

\item{gp}{A \code{\link[grid:gpar]{grid::gpar()}} style to apply}
}
\value{
A list of txt with style attributes
}
\description{
This is a collection of functions to allow styling of text
}
\examples{
fp_txt_italic("Italic text")
}
